remove libc dependency

+++ src/lib.rs
@@ -14,7 +14,7 @@ extern "C" {
 }
 
 #[no_mangle]
-pub extern "C" fn rust_main() -> i32 {
+pub extern "C" fn rust_main() -> i64 {
     unsafe {
         callme();
     }
